Move checks to separate file. Add logging
