+++ core/wormcog.py
@@ -1,5 +1,6 @@
 import asyncio
 import json
+import logging
 import git
 
 import discord
